fix(NotFound): use tokens, remove hardcoded styling, and changed CTA
