--- frontend/src/views/NotFound/index.ts
@@ -1,1 +1,0 @@
-export { default as NotFound } from './NotFound';
